Windows port - CMake option WITH_LIBMARIADB, build with libmariadb from source
As Windows does not have system libraries for MySQL or MariaDB client,
for the sake of simplicity, allow to build libmariadb as external project
during sysbench build.

That is, if WITH_LIBMARIADB is defined, libmariadb will be built from
source https://github.com/mariadb-corporation/mariadb-connector-c,
courtesy of CMake's ExternalProject_Add

libmariadb is chosen because it is easy to build standalone
(as opposed to libmysqlclient) and works well with named pipes,
which libmysqlclient has problems with.

WITH_LIBMARIADB defaults to ON on Windows, as to OFF elsewhere.
It is built as shared library on Windows, as static elsewhere( to
avoid linker warning about mix of C runtimes)