Windows port - export symbols from sysbench executable
The list of symbols is extracted from ffi.cdef sections in internal
lua files into "module definition file" sysbench.def, which is used by
the linker to export symbols.

Alternative to that could be WINDOWS_EXPORT_ALL_SYMBOLS in recent cmake,
but that does not work with LTO/PGO, unfortunately.