Windows port - build and test on CI (github workflows)
The build utilizes vcpkg package manager in order to build with
Postgres driver.

A ZIP sysbench package is built and stored as build artifact,
so if desired, someone can use extract it, and produce official binary
release.

The tests are smoke OLTP tests that run against installed MariaDB and
Postgres.

As Cram unit test do not work on Windows, so this is the only test option
´