Windows port - improve performance by using segment heap
When sysbench is run with many threads, a lot of CPU is spent on memory
allocation and freeing. Profiler shows as much as 80 percent in malloc/free
code, when oltp_read_only is run with 1000 threads.

This does not happen when using SegmentHeap. There, at most 25% time is
spend on allocation and freeing, and the OLTP tps/qps numbers are better.

This patch uses windows application manifest to define heapType=SegmentHeap
for the sysbench executable. If not desired, -DWITH_SEGMENT_HEAP=0 cmake
parameter will disable the segment heap.